Return helpful and clear error code, when Acquisition unit name already exist
